requirements.txt
streamlit>=1.30.0

langchain>=0.2.11
langchain-core>=0.2.23
langchain-community>=0.2.10
langchain-openai>=0.1.17

openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
tiktoken>=0.7.0